+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -   Command `ui` opens webbrowser with Apache Airflow
 -   Move options to config file and restructure
 -   `MLflow` enabled only if `kedro-mlflow` present in dependencies and configuration is in place 
+-   Command `init` to initialize configuration for the plugin in kedro project
 
 ## [0.1.2] - 2021-03-24